Add mqtt output plugin to logstash indexers
This commit adds support for configuring the mqtt logstash output
plugin. This will enable us to push events from our logstash workers
to firehose.o.o. The plugin can be found here:

https://github.com/kompa3/logstash-output-mqtt (althought it's README
is from a template and you have to look at the inline docs for how
to use it)

and the gem is published on rubygems.org here:

https://rubygems.org/gems/logstash-output-mqtt/

Change-Id: I7255f6c256ab3e3ca33caf69b71bf4ffab02c1bb